Time: 0 ms (100%), Space: 15.1 MB (61.23%) - LeetHub

+++ 2714-left-and-right-sum-differences/2714-left-and-right-sum-differences.cpp
@@ -19,7 +19,7 @@ public:
         vector<int> answer;
         answer.reserve(numsSize);
 
-        for (short i = 0; i < numsSize; ++i)
+        for (int i = 0; i < numsSize; ++i)
             answer.push_back(abs(leftSum[i] - rightSum[numsSize - 1 - i]));
 
         return answer;
